pyproject.toml
[build-system]
requires = [
    "setuptools>=60",
    "setuptools-scm>=8.0"]

[project]
name = "scdemon"
version = "0.0.4"
description = "Single cell decomposition into module networks"
authors = [{name = "Benjamin James", email = "benjames@mit.edu"},
           {name = "Carles Boix", email = "cboix@mit.edu"}]
license = {file = "LICENSE"}
requires-python = ">=3.7"
dependencies = [
    "numpy", "pandas", 
    "scipy", "igraph", 
    "umap-learn", "leidenalg", 
    "scanpy", "seaborn", "matplotlib"
]

[project.urls]
Documentation = "https://scdemon.readthedocs.io/"
Source = "https://github.com/kellislab/scdemon"
Home-page = "https://github.com/kellislab/scdemon"

[project.optional-dependencies]
test = ["adjustText", "gprofiler-official", "anndata"]

[tool.setuptools]
packages = ["scdemon", 
         "scdemon.utils", "scdemon.plotting",
         "scdemon.graph", "scdemon.data"]

[tool.coverage.run]
branch = true

